+++ deploy/terraform/examples/kubernetes/outputs.tf
@@ -1,0 +1,7 @@
+output "release" {
+  value = {
+    name      = module.dmarq.release_name
+    namespace = module.dmarq.namespace
+    status    = module.dmarq.status
+  }
+}